dm: core: Change driver data to be uintptr_t
The driver data field is often used to store a pointer, but is defined
as an unsigned long. This will fail for platforms where uintptr_t is a
different length from unsigned long, such as CHERI. This patch also
changes some of the drivers that are casting to ulong instead of
uintptr_t.

+++ drivers/net/rswitch.c
@@ -1211,7 +1211,7 @@ static int rswitch_bind(struct udevice *parent)
 
 		ret = device_bind_with_driver_data(parent, drv,
 						   ofnode_get_name(node),
-						   (ulong)priv, node, &dev);
+						   (uintptr_t)priv, node, &dev);
 		if (ret)
 			return ret;
 	}
